+++ server/routes/users.js
@@ -4,7 +4,8 @@ const router = require('express').Router()
 const webtoken = require('jsonwebtoken')
 const TOKEN_SECRET = process.env.TOKEN_SECRET
 const STATUS_CODES = require('../statusCodes')
-const { recalculateSessionAverages, recalculateInteractionAverages, LIKE} = require('../constants')
+const { LIKE } = require('../constants')
+const { recalculateSessionAverages, recalculateInteractionAverages } = require('../sessionUtils')
 
 const prisma = new PrismaClient()
